Options.Applicative.BashCompletion

Description

You don't need to import this module to enable bash completion.

See the wiki for more information on bash completion.

Documentation

bashCompletionParser :: ParserInfo a -> ParserPrefs -> Parser CompletionResult