Options.Applicative.Common

Option parsers

A Parser is composed of a list of options. Several kinds of options are supported:

Flags: simple no-argument options. When a flag is encountered on the command line, its value is returned.

Options: options with an argument. An option can define a reader, which converts its argument from String to the desired value, or throws a parse error if the argument does not validate correctly.

Arguments: positional arguments, validated in the same way as option arguments.

Commands. A command defines a completely independent sub-parser. When a command is encountered, the whole command line is passed to the corresponding parser.

data Parser a

A Parser a is an option parser returning a value of type a.

liftOpt :: Option a -> Parser a

Create a parser composed of a single option.

showOption :: OptName -> String

Program descriptions

data ParserInfo a

A full description for a runnable Parser for a program.

Constructors

ParserInfo

Fields

infoParser :: Parser a

the option parser for the program

infoFullDesc :: Bool

whether the help text should contain full documentation

infoProgDesc :: Chunk Doc

brief parser description

infoHeader :: Chunk Doc

header of the full parser description

infoFooter :: Chunk Doc

footer of the full parser description

infoFailureCode :: Int

exit code for a parser failure

infoPolicy :: ArgPolicy

allow regular options and flags to occur after arguments (default: InterspersePolicy)

data ParserPrefs

Global preferences for a top-level Parser.

Constructors

ParserPrefs

Fields

prefMultiSuffix :: String

metavar suffix for multiple options

prefDisambiguate :: Bool

automatically disambiguate abbreviations (default: False)

prefShowHelpOnError :: Bool

always show help text on parse errors (default: False)

prefShowHelpOnEmpty :: Bool

show the help text for a command or subcommand if it fails with no input (default: False)

prefBacktrack :: Backtracking

backtrack to parent parser when a subcommand fails (default: Backtrack)

prefColumns :: Int

number of columns in the terminal, used to format the help page (default: 80)

prefHelpLongEquals :: Bool

when displaying long names in usage and help, use an '=' sign for long names, rather than a single space (default: False)

prefHelpShowGlobal :: Bool

when displaying subparsers' usage help, show parent options under a "global options" section (default: False)

prefTabulateFill :: Int

Indentation width for tables

Running parsers

runParserInfo :: MonadP m => ParserInfo a -> Args -> m a

runParserFully :: MonadP m => ArgPolicy -> Parser a -> Args -> m a

runParserStep :: MonadP m => ArgPolicy -> Parser a -> String -> Args -> m (Maybe (Parser a), Args)

runParser :: MonadP m => ArgPolicy -> IsCmdStart -> Parser a -> Args -> m (a, Args)

Apply a Parser to a command line, and return a result and leftover arguments. This function returns an error if any parsing error occurs, or if any options are missing and don't have a default value.

evalParser :: Parser a -> Maybe a

The default value of a Parser. This function returns an error if any of the options don't have a default value.

Low-level utilities

mapParser :: (forall x. ArgumentReachability -> Option x -> b) -> Parser a -> [b]

Map a polymorphic function over all the options of a parser, and collect the results in a list.

treeMapParser :: (forall x. ArgumentReachability -> Option x -> b) -> Parser a -> OptTree b

Like mapParser, but collect the results in a tree structure.

optionNames :: OptReader a -> [OptName]