Options.Applicative.Types

Documentation

data ParseError

Constructors

ErrorMsg String

InfoMsg String

ShowHelpText (Maybe String)

UnknownError

MissingError IsCmdStart SomeParser

ExpectsArgError String

UnexpectedError String SomeParser

data ParserInfo a

A full description for a runnable Parser for a program.

Constructors

ParserInfo

Fields

infoParser :: Parser a

the option parser for the program

infoFullDesc :: Bool

whether the help text should contain full documentation

infoProgDesc :: Chunk Doc

brief parser description

infoHeader :: Chunk Doc

header of the full parser description

infoFooter :: Chunk Doc

footer of the full parser description

infoFailureCode :: Int

exit code for a parser failure

infoPolicy :: ArgPolicy

allow regular options and flags to occur after arguments (default: InterspersePolicy)

data ParserPrefs

Global preferences for a top-level Parser.

Constructors

ParserPrefs

Fields

prefMultiSuffix :: String

metavar suffix for multiple options

prefDisambiguate :: Bool

automatically disambiguate abbreviations (default: False)

prefShowHelpOnError :: Bool

always show help text on parse errors (default: False)

prefShowHelpOnEmpty :: Bool

show the help text for a command or subcommand if it fails with no input (default: False)

prefBacktrack :: Backtracking

backtrack to parent parser when a subcommand fails (default: Backtrack)

prefColumns :: Int

number of columns in the terminal, used to format the help page (default: 80)

prefHelpLongEquals :: Bool

when displaying long names in usage and help, use an '=' sign for long names, rather than a single space (default: False)

prefHelpShowGlobal :: Bool

when displaying subparsers' usage help, show parent options under a "global options" section (default: False)

prefTabulateFill :: Int

Indentation width for tables

data Option a

A single option of a parser.

Constructors

Option

Fields

optMain :: OptReader a

reader for this option

optProps :: OptProperties

properties of this option

data OptName

Constructors

OptShort !Char

OptLong !String

isShortName :: OptName -> Bool

isLongName :: OptName -> Bool

data OptReader a

An OptReader defines whether an option matches an command line argument.

Constructors

OptReader [OptName] (CReader a) (String -> ParseError)
option reader

FlagReader [OptName] !a
flag reader

ArgReader (CReader a)
argument reader

CmdReader (Maybe String) [(String, ParserInfo a)]
command reader

data OptProperties

Specification for an individual parser option.

Constructors

OptProperties

Fields

propVisibility :: OptVisibility

whether this flag is shown in the brief description

propHelp :: Chunk Doc

help text for this option

propMetaVar :: String

metavariable for this option

propShowDefault :: Maybe String

what to show in the help text as the default

propShowGlobal :: Bool

whether the option is presented in global options text

propDescMod :: Maybe (Doc -> Doc)

a function to run over the brief description

data OptVisibility

Visibility of an option in the help text.

Constructors

Internal
does not appear in the help text at all

Hidden
only visible in the full description

Visible
visible both in the full and brief descriptions

data Backtracking

Constructors

Backtrack

NoBacktrack

SubparserInline

newtype ReadM a

A newtype over 'ReaderT String Except', used by option readers.

Constructors

ReadM

Fields

unReadM :: ReaderT String (Except ParseError) a

readerAsk :: ReadM String

Return the value being read.

readerAbort :: ParseError -> ReadM a

Abort option reader by exiting with a ParseError.

readerError :: String -> ReadM a

Abort option reader by exiting with an error message.

data CReader a

Constructors

CReader

Fields

crCompleter :: Completer

crReader :: ReadM a

data Parser a

A Parser a is an option parser returning a value of type a.

Constructors

NilP (Maybe a)

OptP (Option a)

forall x. MultP (Parser (x -> a)) (Parser x)

AltP (Parser a) (Parser a)

forall x. BindP (Parser x) (x -> Parser a)

newtype ParserM r

Constructors

ParserM

Fields

runParserM :: forall x. (r -> Parser x) -> Parser x

newtype Completer

A shell complete function.

Constructors

Completer

Fields

runCompleter :: String -> IO [String]

mkCompleter :: (String -> IO [String]) -> Completer

Smart constructor for a Completer

newtype CompletionResult

Constructors

CompletionResult

Fields

execCompletion :: String -> IO String

newtype ParserFailure h

Constructors

ParserFailure

Fields

execFailure :: String -> (h, ExitCode, Int)

data ParserResult a

Result of execParserPure.

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult

overFailure :: (ParserHelp -> ParserHelp) -> ParserResult a -> ParserResult a

type Args = [String]

data ArgPolicy

Policy for how to handle options within the parse

Constructors

Intersperse
The default policy, options and arguments can be interspersed. A `--` option can be passed to ensure all following commands are treated as arguments.

NoIntersperse
Options must all come before arguments, once a single positional argument or subcommand is parsed, all remaining arguments are treated as positionals. A `--` option can be passed if the first positional one needs starts with -.

AllPositionals
No options are parsed at all, all arguments are treated as positionals. Is the policy used after `--` is encountered.

ForwardOptions
Options and arguments can be interspersed, but if a given option is not found, it is treated as a positional argument. This is sometimes useful if one is passing through most options to another tool, but are supplying just a few of their own options.

newtype ArgumentReachability

Constructors

ArgumentReachability

Fields

argumentIsUnreachable :: Bool

If the result is a positional, if it can't be accessed in the current parser position ( first arg )

data AltNodeType

This type encapsulates whether an AltNode of an OptTree should be displayed with brackets around it.

Constructors

MarkDefault

NoDefault

data OptTree a

Constructors

Leaf a

MultNode [OptTree a]

AltNode AltNodeType [OptTree a]

BindNode (OptTree a)

data ParserHelp

Constructors

ParserHelp

Fields

helpError :: Chunk Doc

helpSuggestions :: Chunk Doc

helpHeader :: Chunk Doc

helpUsage :: Chunk Doc

helpDescription :: Chunk Doc

helpBody :: Chunk Doc

helpGlobals :: Chunk Doc

helpFooter :: Chunk Doc

data SomeParser

Constructors

forall a. SomeParser (Parser a)

data Context

Subparser context, containing the name of the subparser and its parser info. Used by parserFailure to display relevant usage information when parsing inside a subparser fails.

Constructors

forall a. Context String (ParserInfo a)

data IsCmdStart

Constructors

CmdStart

CmdCont

fromM :: ParserM a -> Parser a

oneM :: Parser a -> ParserM a

manyM :: Parser a -> ParserM [a]

someM :: Parser a -> ParserM [a]

filterOptional :: OptTree a -> OptTree a

optVisibility :: Option a -> OptVisibility

optMetaVar :: Option a -> String

optHelp :: Option a -> Chunk Doc

optShowDefault :: Option a -> Maybe String

optDescMod :: Option a -> Maybe (Doc -> Doc)