Options.Applicative.Utils

Documentation

(<+>) :: String -> String -> String

Concatenate two strings with a space in the middle.

vcat :: [String] -> String

Concatenate strings vertically with empty lines in between.

tabulate :: [(String, String)] -> [String]

Display pairs of strings in a table.

pad :: Int -> String -> String

Pad a string to a fixed size with whitespace.