Options.Applicative.Builder.Internal

Internals

data Mod f a

An option modifier.

Option modifiers are values that represent a modification of the properties of an option.

The type parameter a is the return type of the option, while f is a record containing its properties (e.g. OptionFields for regular options, FlagFields for flags, etc...).

An option modifier consists of 3 elements:

A field modifier, of the form f a -> f a. These are essentially (compositions of) setters for some of the properties supported by f.

An optional default value and function to display it.

A property modifier, of the form OptProperties -> OptProperties. This is just like the field modifier, but for properties applicable to any option.

Modifiers are instances of Monoid, and can be composed as such.

You rarely need to deal with modifiers directly, as most of the times it is sufficient to pass them to builders (such as strOption or flag) to create options (see Builder).

Constructors

Mod (f a -> f a) (DefaultProp a) (OptProperties -> OptProperties)

class HasName f where

Methods

name :: OptName -> f a -> f a

class HasCompleter f where

Methods

modCompleter :: (Completer -> Completer) -> f a -> f a

data OptionFields a

Constructors

OptionFields

Fields

optNames :: [OptName]

optCompleter :: Completer

optReader :: String -> ReadM a

optNoArgError :: ParseError

data FlagFields a

Constructors

FlagFields

Fields

flagNames :: [OptName]

flagActive :: a

data CommandFields a

Constructors

CommandFields

Fields

cmdCommands :: [(String, ParserInfo a)]

data ArgumentFields a

Constructors

ArgumentFields

Fields

argCompleter :: Completer

data DefaultProp a

Constructors

DefaultProp (Maybe a) (Maybe (a -> String))

optionMod :: (OptProperties -> OptProperties) -> Mod f a

fieldMod :: (f a -> f a) -> Mod f a

baseProps :: OptProperties

Base default properties.

mkCommand :: Mod CommandFields a -> ([String], String -> Maybe (ParserInfo a))

mkParser :: DefaultProp a -> (OptProperties -> OptProperties) -> OptReader a -> Parser a

mkOption :: DefaultProp a -> (OptProperties -> OptProperties) -> OptReader a -> Option a

mkProps :: DefaultProp a -> (OptProperties -> OptProperties) -> OptProperties

internal :: Mod f a

Hide this option from the help text