Options.Applicative.Extra

Extra parser utilities

This module contains high-level functions to run parsers.

helper :: Parser (a -> a)

A hidden "helper" option which always fails.

A common usage pattern is to apply this applicatively when creating a ParserInfo

opts :: ParserInfo Sample
opts = info (sample <**> helper) mempty

hsubparser :: Mod CommandFields a -> Parser a

Builder for a command parser with a "helper" option attached. Used in the same way as subparser, but includes a "--help|-h" inside the subcommand.

execParser :: ParserInfo a -> IO a

Run a program description.

Parse command line arguments. Display help text and exit if any parse error occurs.

execParserMaybe :: ParserInfo a -> [String] -> Maybe a

Deprecated: Use execParserPure together with getParseResult instead

Run a program description in pure code.

This function behaves like execParser, but can be called from pure code. Note that, in case of errors, no message is displayed, and this function simply returns Nothing.

If you need to keep track of error messages, use execParserPure instead.

customExecParser :: ParserPrefs -> ParserInfo a -> IO a

Run a program description with custom preferences.

customExecParserMaybe :: ParserPrefs -> ParserInfo a -> [String] -> Maybe a

Deprecated: Use execParserPure together with getParseResult instead

Run a program description with custom preferences in pure code.

See execParserMaybe for details.

execParserPure

Arguments

:: ParserPrefs
Global preferences for this parser

-> ParserInfo a
Description of the program to run

-> [String]
Program arguments

-> ParserResult a

The most general way to run a program description in pure code.

getParseResult :: ParserResult a -> Maybe a

Extract the actual result from a ParserResult value.

This function returns Nothing in case of errors. Possible error messages or completion actions are simply discarded.

If you want to display error messages and invoke completion actions appropriately, use handleParseResult instead.

handleParseResult :: ParserResult a -> IO a

Handle ParserResult.

parserFailure :: ParserPrefs -> ParserInfo a -> ParseError -> [Context] -> ParserFailure ParserHelp

Generate a ParserFailure from a ParseError in a given Context.

This function can be used, for example, to show the help text for a parser:

handleParseResult . Failure $ parserFailure pprefs pinfo ShowHelpText mempty

renderFailure :: ParserFailure ParserHelp -> String -> (String, ExitCode)

newtype ParserFailure h

Constructors

ParserFailure

Fields

execFailure :: String -> (h, ExitCode, Int)

overFailure :: (ParserHelp -> ParserHelp) -> ParserResult a -> ParserResult a

data ParserResult a

Result of execParserPure.

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult

data ParserPrefs

Global preferences for a top-level Parser.

Constructors

ParserPrefs

Fields

prefMultiSuffix :: String

metavar suffix for multiple options

prefDisambiguate :: Bool

automatically disambiguate abbreviations (default: False)

prefShowHelpOnError :: Bool

always show help text on parse errors (default: False)

prefShowHelpOnEmpty :: Bool

show the help text for a command or subcommand if it fails with no input (default: False)

prefBacktrack :: Bool

backtrack to parent parser when a subcommand fails (default: True)

prefColumns :: Int

number of columns in the terminal, used to format the help page (default: 80)

newtype CompletionResult

Constructors

CompletionResult

Fields

execCompletion :: String -> IO String