Options.Applicative.Extra

Extra parser utilities

This module contains high-level functions to run parsers.

helper :: Parser (a -> a)

A hidden "helper" option which always fails.

hsubparser :: Mod CommandFields a -> Parser a

execParser :: ParserInfo a -> IO a

Run a program description.

Parse command line arguments. Display help text and exit if any parse error occurs.

execParserMaybe :: ParserInfo a -> [String] -> Maybe a

Run a program description in pure code.

This function behaves like execParser, but can be called from pure code. Note that, in case of errors, no message is displayed, and this function simply returns Nothing.

If you need to keep track of error messages, use execParserPure instead.

customExecParser :: ParserPrefs -> ParserInfo a -> IO a

Run a program description with custom preferences.

customExecParserMaybe :: ParserPrefs -> ParserInfo a -> [String] -> Maybe a

Run a program description with custom preferences in pure code.

See execParserMaybe for details.

execParserPure

Arguments

:: ParserPrefs
Global preferences for this parser

-> ParserInfo a
Description of the program to run

-> [String]
Program arguments

-> Either ParserFailure a

The most general way to run a program description in pure code.

usage :: ParserPrefs -> Parser a -> String -> String

Generate option summary.

data ParserFailure

Result after a parse error.

Constructors

ParserFailure

Fields

errMessage :: String -> IO String

Function which takes the program name as input and returns an error message

errExitCode :: ExitCode

Exit code to use for this error