Options.Applicative.Extra

Extra parser utilities

This module contains high-level functions to run parsers.

helper :: Parser (a -> a)

A hidden "helper" option which always fails.

execParser :: ParserInfo a -> IO a

Run a program description.

Parse command line arguments. Display help text and exit if any parse error occurs.

execParserPure

Arguments

:: ParserPrefs
Global preferences for this parser

-> ParserInfo a
Description of the program to run

-> [String]
Program arguments

-> Either ParserFailure a

A pure version execParser.

customExecParser :: ParserPrefs -> ParserInfo a -> IO a

Run a program description with custom preferences.

usage :: ParserPrefs -> Parser a -> String -> String

Generate option summary.

data ParserFailure

Result after a parse error.

Constructors

ParserFailure

Fields

errMessage :: String -> IO String

Function which takes the program name as input and returns an error message

errExitCode :: ExitCode

Exit code to use for this error