Options.Applicative.Help.Pretty

Documentation

hPutDoc :: Handle -> Doc -> IO ()

(hPutDoc handle doc) pretty prints document doc to the file handle handle with a page width of 80 characters and a ribbon width of 32 characters.

main = do{ handle <- openFile "MyFile" WriteMode
         ; hPutDoc handle (vcat (map text
                           ["vertical","text"]))
         ; hClose handle
         }

Any ANSI colorisation in doc will be output.

putDoc :: Doc -> IO ()

The action (putDoc doc) pretty prints document doc to the standard output, with a page width of 80 characters and a ribbon width of 32 characters.

main :: IO ()
main = do{ putDoc (text "hello" <+> text "world") }

Which would output

hello world

Any ANSI colorisation in doc will be output.

displayIO :: Handle -> SimpleDoc -> IO ()

(displayIO handle simpleDoc) writes simpleDoc to the file handle handle. This function is used for example by hPutDoc:

hPutDoc handle doc  = displayIO handle (renderPretty 0.4 80 doc)

Any ANSI colorisation in simpleDoc will be output.

renderCompact :: Doc -> SimpleDoc

(renderCompact x) renders document x without adding any indentation. Since no 'pretty' printing is involved, this renderer is very fast. The resulting output contains fewer characters than a pretty printed version and can be used for output that is read by other programs.

This rendering function does not add any colorisation information.

renderSmart :: Float -> Int -> Doc -> SimpleDoc

A slightly smarter rendering algorithm with more lookahead. It provides provide earlier breaking on deeply nested structures For example, consider this python-ish pseudocode: fun(fun(fun(fun(fun([abcdefg, abcdefg]))))) If we put a softbreak (+ nesting 2) after each open parenthesis, and align the elements of the list to match the opening brackets, this will render with renderPretty and a page width of 20 as: fun(fun(fun(fun(fun([ | abcdef, | abcdef, ] ))))) | Where the 20c. boundary has been marked with |. Because renderPretty only uses one-line lookahead, it sees that the first line fits, and is stuck putting the second and third lines after the 20-c mark. In contrast, renderSmart will continue to check that the potential document up to the end of the indentation level. Thus, it will format the document as:

fun(                |
  fun(              |
    fun(            |
      fun(          |
        fun([       |
              abcdef,
              abcdef,
            ]       |
  )))))             |

Which fits within the 20c. boundary.

plain :: Doc -> Doc

Removes all colorisation, emboldening and underlining from a document

deunderline :: Doc -> Doc

Displays a document with no underlining

underline :: Doc -> Doc

Displays a document with underlining

debold :: Doc -> Doc

Displays a document in the normal font weight

bold :: Doc -> Doc

Displays a document in a heavier font weight

ondullwhite :: Doc -> Doc

Displays a document with the dull white backcolor

onwhite :: Doc -> Doc

Displays a document with the white backcolor

ondullcyan :: Doc -> Doc

Displays a document with the dull cyan backcolor

oncyan :: Doc -> Doc

Displays a document with the cyan backcolor

ondullmagenta :: Doc -> Doc

Displays a document with the dull magenta backcolor

onmagenta :: Doc -> Doc

Displays a document with the magenta backcolor

ondullblue :: Doc -> Doc

Displays a document with the dull blue backcolor

onblue :: Doc -> Doc

Displays a document with the blue backcolor

ondullyellow :: Doc -> Doc

Displays a document with the dull yellow backcolor

onyellow :: Doc -> Doc

Displays a document with the yellow backcolor

ondullgreen :: Doc -> Doc

Displays a document with the dull green backcolor

ongreen :: Doc -> Doc

Displays a document with the green backcolor

ondullred :: Doc -> Doc

Displays a document with the dull red backcolor

onred :: Doc -> Doc

Displays a document with the red backcolor

ondullblack :: Doc -> Doc

Displays a document with the dull black backcolor

onblack :: Doc -> Doc

Displays a document with the black backcolor

dullwhite :: Doc -> Doc

Displays a document with the dull white forecolor

white :: Doc -> Doc

Displays a document with the white forecolor

dullcyan :: Doc -> Doc

Displays a document with the dull cyan forecolor

cyan :: Doc -> Doc

Displays a document with the cyan forecolor

dullmagenta :: Doc -> Doc

Displays a document with the dull magenta forecolor

magenta :: Doc -> Doc

Displays a document with the magenta forecolor

dullblue :: Doc -> Doc

Displays a document with the dull blue forecolor

blue :: Doc -> Doc

Displays a document with the blue forecolor

dullyellow :: Doc -> Doc

Displays a document with the dull yellow forecolor

yellow :: Doc -> Doc

Displays a document with the yellow forecolor

dullgreen :: Doc -> Doc

Displays a document with the dull green forecolor

green :: Doc -> Doc

Displays a document with the green forecolor

dullred :: Doc -> Doc

Displays a document with the dull red forecolor

red :: Doc -> Doc

Displays a document with the red forecolor

dullblack :: Doc -> Doc

Displays a document with the dull black forecolor

black :: Doc -> Doc

Displays a document with the black forecolor

flatAlt :: Doc -> Doc -> Doc

A document that is normally rendered as the first argument, but when flattened, is rendered as the second document.

group :: Doc -> Doc

The group combinator is used to specify alternative layouts. The document (group x) undoes all line breaks in document x. The resulting line is added to the current line if that fits the page. Otherwise, the document x is rendered without any changes.

nesting :: (Int -> Doc) -> Doc

column :: (Int -> Doc) -> Doc

nest :: Int -> Doc -> Doc

The document (nest i x) renders document x with the current indentation level increased by i (See also hang, align and indent).

nest 2 (text "hello" <$> text "world") <$> text "!"

outputs as:

hello
  world
!

hardline :: Doc

A linebreak that will never be flattened; it is guaranteed to render as a newline.

linebreak :: Doc

The linebreak document advances to the next line and indents to the current nesting level. Document linebreak behaves like empty if the line break is undone by group.

line :: Doc

The line document advances to the next line and indents to the current nesting level. Document line behaves like (text " ") if the line break is undone by group.

text :: String -> Doc

The document (text s) contains the literal string s. The string shouldn't contain any newline ('n') characters. If the string contains newline characters, the function string should be used.

char :: Char -> Doc

The document (char c) contains the literal character c. The character shouldn't be a newline ('n'), the function line should be used for line breaks.

empty :: Doc

The empty document is, indeed, empty. Although empty has no content, it does have a 'height' of 1 and behaves exactly like (text "") (and is therefore not a unit of <$>).

align :: Doc -> Doc

The document (align x) renders document x with the nesting level set to the current column. It is used for example to implement hang.

As an example, we will put a document right above another one, regardless of the current nesting level:

x $$ y  = align (x <$> y)

test    = text "hi" <+> (text "nice" $$ text "world")

which will be layed out as:

hi nice
   world

hang :: Int -> Doc -> Doc

The hang combinator implements hanging indentation. The document (hang i x) renders document x with a nesting level set to the current column plus i. The following example uses hanging indentation for some text:

test  = hang 4 (fillSep (map text
        (words "the hang combinator indents these words !")))

Which lays out on a page with a width of 20 characters as:

the hang combinator
    indents these
    words !

The hang combinator is implemented as:

hang i x  = align (nest i x)

width :: Doc -> (Int -> Doc) -> Doc

fill :: Int -> Doc -> Doc

The document (fill i x) renders document x. It than appends spaces until the width is equal to i. If the width of x is already larger, nothing is appended. This combinator is quite useful in practice to output a list of bindings. The following example demonstrates this.

types  = [("empty","Doc")
         ,("nest","Int -> Doc -> Doc")
         ,("linebreak","Doc")]

ptype (name,tp)
       = fill 6 (text name) <+> text "::" <+> text tp

test   = text "let" <+> align (vcat (map ptype types))

Which is layed out as:

let empty  :: Doc
    nest   :: Int -> Doc -> Doc
    linebreak :: Doc

fillBreak :: Int -> Doc -> Doc

The document (fillBreak i x) first renders document x. It than appends spaces until the width is equal to i. If the width of x is already larger than i, the nesting level is increased by i and a line is appended. When we redefine ptype in the previous example to use fillBreak, we get a useful variation of the previous output:

ptype (name,tp)
       = fillBreak 6 (text name) <+> text "::" <+> text tp

The output will now be:

let empty  :: Doc
    nest   :: Int -> Doc -> Doc
    linebreak
           :: Doc

rational :: Rational -> Doc

The document (rational r) shows the literal rational r using text.

double :: Double -> Doc

The document (double d) shows the literal double d using text.

float :: Float -> Doc

The document (float f) shows the literal float f using text.

integer :: Integer -> Doc

The document (integer i) shows the literal integer i using text.

int :: Int -> Doc

The document (int i) shows the literal integer i using text.

bool :: Bool -> Doc

The document (bool b) shows the literal bool b using text.

string :: String -> Doc

The document (string s) concatenates all characters in s using line for newline characters and char for all other characters. It is used instead of text whenever the text contains newline characters.

equals :: Doc

The document equals contains an equal sign, "=".

backslash :: Doc

The document backslash contains a back slash, "\".

dot :: Doc

The document dot contains a single dot, ".".

space :: Doc

The document space contains a single space, " ".

x <+> y   = x <> space <> y

comma :: Doc

The document comma contains a comma, ",".

colon :: Doc

The document colon contains a colon, ":".

semi :: Doc

The document semi contains a semicolon, ";".

dquote :: Doc

The document dquote contains a double quote, '"'.

squote :: Doc

The document squote contains a single quote, "'".

rbracket :: Doc

The document rbracket contains a right square bracket, "]".

lbracket :: Doc

The document lbracket contains a left square bracket, "[".

rbrace :: Doc

The document rbrace contains a right brace, "}".

lbrace :: Doc

The document lbrace contains a left brace, "{".

rangle :: Doc

The document rangle contains a right angle, ">".

langle :: Doc

The document langle contains a left angle, "<".

rparen :: Doc

The document rparen contains a right parenthesis, ")".

lparen :: Doc

The document lparen contains a left parenthesis, "(".

enclose :: Doc -> Doc -> Doc -> Doc

The document (enclose l r x) encloses document x between documents l and r using (<>).

enclose l r x   = l <> x <> r

brackets :: Doc -> Doc

Document (brackets x) encloses document x in square brackets, "[" and "]".

angles :: Doc -> Doc

Document (angles x) encloses document x in angles, "<" and ">".

parens :: Doc -> Doc

Document (parens x) encloses document x in parenthesis, "(" and ")".

braces :: Doc -> Doc

Document (braces x) encloses document x in braces, "{" and "}".

dquotes :: Doc -> Doc

Document (dquotes x) encloses document x with double quotes '"'.

squotes :: Doc -> Doc

Document (squotes x) encloses document x with single quotes "'".

softbreak :: Doc

The document softbreak behaves like empty if the resulting output fits the page, otherwise it behaves like line.

softbreak  = group linebreak

softline :: Doc

The document softline behaves like space if the resulting output fits the page, otherwise it behaves like line.

softline = group line

(<$$>) :: Doc -> Doc -> Doc infixr 5

The document (x <$$> y) concatenates document x and y with a linebreak in between. (infixr 5)

(<//>) :: Doc -> Doc -> Doc infixr 5

The document (x <//> y) concatenates document x and y with a softbreak in between. This effectively puts x and y either right next to each other or underneath each other. (infixr 5)

(</>) :: Doc -> Doc -> Doc infixr 5

The document (x </> y) concatenates document x and y with a softline in between. This effectively puts x and y either next to each other (with a space in between) or underneath each other. (infixr 5)

(<+>) :: Doc -> Doc -> Doc infixr 6

The document (x <+> y) concatenates document x and y with a space in between. (infixr 6)

vcat :: [Doc] -> Doc

The document (vcat xs) concatenates all documents xs vertically with (<$$>). If a group undoes the line breaks inserted by vcat, all documents are directly concatenated.

hcat :: [Doc] -> Doc

The document (hcat xs) concatenates all documents xs horizontally with (<>).

fillCat :: [Doc] -> Doc

The document (fillCat xs) concatenates documents xs horizontally with (<>) as long as its fits the page, than inserts a linebreak and continues doing that for all documents in xs.

fillCat xs  = foldr (<//>) empty xs

cat :: [Doc] -> Doc

The document (cat xs) concatenates all documents xs either horizontally with (<>), if it fits the page, or vertically with (<$$>).

cat xs  = group (vcat xs)

vsep :: [Doc] -> Doc

The document (vsep xs) concatenates all documents xs vertically with (<$>). If a group undoes the line breaks inserted by vsep, all documents are separated with a space.

someText = map text (words ("text to lay out"))

test     = text "some" <+> vsep someText

This is layed out as:

some text
to
lay
out

The align combinator can be used to align the documents under their first element

test     = text "some" <+> align (vsep someText)

Which is printed as:

some text
     to
     lay
     out

hsep :: [Doc] -> Doc

The document (hsep xs) concatenates all documents xs horizontally with (<+>).

fillSep :: [Doc] -> Doc

The document (fillSep xs) concatenates documents xs horizontally with (<+>) as long as its fits the page, than inserts a line and continues doing that for all documents in xs.

fillSep xs  = foldr (</>) empty xs

sep :: [Doc] -> Doc

The document (sep xs) concatenates all documents xs either horizontally with (<+>), if it fits the page, or vertically with (<$>).

sep xs  = group (vsep xs)

punctuate :: Doc -> [Doc] -> [Doc]

(punctuate p xs) concatenates all documents in xs with document p except for the last document.

someText = map text ["words","in","a","tuple"]
test     = parens (align (cat (punctuate comma someText)))

This is layed out on a page width of 20 as:

(words,in,a,tuple)

But when the page width is 15, it is layed out as:

(words,
 in,
 a,
 tuple)

(If you want put the commas in front of their elements instead of at the end, you should use tupled or, in general, encloseSep.)

encloseSep :: Doc -> Doc -> Doc -> [Doc] -> Doc

The document (encloseSep l r sep xs) concatenates the documents xs separated by sep and encloses the resulting document by l and r. The documents are rendered horizontally if that fits the page. Otherwise they are aligned vertically. All separators are put in front of the elements. For example, the combinator list can be defined with encloseSep:

list xs = encloseSep lbracket rbracket comma xs
test    = text "list" <+> (list (map int [10,200,3000]))

Which is layed out with a page width of 20 as:

list [10,200,3000]

But when the page width is 15, it is layed out as:

list [10
     ,200
     ,3000]

semiBraces :: [Doc] -> Doc

The document (semiBraces xs) separates the documents xs with semicolons and encloses them in braces. The documents are rendered horizontally if that fits the page. Otherwise they are aligned vertically. All semicolons are put in front of the elements.

tupled :: [Doc] -> Doc

The document (tupled xs) comma separates the documents xs and encloses them in parenthesis. The documents are rendered horizontally if that fits the page. Otherwise they are aligned vertically. All comma separators are put in front of the elements.

list :: [Doc] -> Doc

The document (list xs) comma separates the documents xs and encloses them in square brackets. The documents are rendered horizontally if that fits the page. Otherwise they are aligned vertically. All comma separators are put in front of the elements.

class Pretty a where

The member prettyList is only used to define the instance Pretty a => Pretty [a]. In normal circumstances only the pretty function is used.

Minimal complete definition

pretty

Methods

pretty :: a -> Doc

prettyList :: [a] -> Doc

data SimpleDoc

The data type SimpleDoc represents rendered documents and is used by the display functions.

Whereas values of the data type Doc represent non-empty sets of possible renderings of a document, values of the data type SimpleDoc represent single renderings of a document.

The Int in SText contains the length of the string. The Int in SLine contains the indentation for that line. The library provides two default display functions displayS and displayIO. You can provide your own display function by writing a function from a SimpleDoc to your own output format.

Constructors

SFail

SEmpty

SChar Char SimpleDoc

SText !Int String SimpleDoc

SLine !Int SimpleDoc

SSGR [SGR] SimpleDoc

type Doc = Doc

indent :: Int -> Doc -> Doc

renderPretty :: Float -> Int -> Doc -> SimpleDoc

displayS :: SimpleDoc -> ShowS

(.$.) :: Doc -> Doc -> Doc

groupOrNestLine :: Doc -> Doc

Render flattened text on this line, or start a new line before rendering any text.

This will also nest subsequent lines in the group.

altSep :: Doc -> Doc -> Doc

Separate items in an alternative with a pipe.

If the first document and the pipe don't fit on the line, then mandatorily flow the next entry onto the following line.

The (//) softbreak ensures that if the document does fit on the line, there is at least a space, but it's possible for y to still appear on the next line.

hangAtIfOver :: Int -> Int -> Doc -> Doc

Printer hacks to get nice indentation for long commands and subcommands.

If we're starting this section over the desired width (usually 1/3 of the ribbon), then we will make a line break, indent all of the usage, and go.

The ifAtRoot is an interesting clause. If this whole operation is put under a group then the linebreak will disappear; then item d will therefore not be at the starting column, and it won't be indented more.