Options.Applicative.Types

Documentation

data ParserInfo a

A full description for a runnable Parser for a program.

Constructors

ParserInfo

Fields

infoParser :: Parser a

the option parser for the program

infoFullDesc :: Bool

whether the help text should contain full documentation

infoProgDesc :: String

brief parser description

infoHeader :: String

header of the full parser description

infoFooter :: String

footer of the full parser description

infoFailureCode :: Int

exit code for a parser failure

data ParserPrefs

Global preferences for a top-level Parser.

Constructors

ParserPrefs

Fields

prefMultiSuffix :: String

metavar suffix for multiple options

prefDisambiguate :: Bool

automatically disambiguate abbreviations

data Option a

A single option of a parser.

Constructors

Option

Fields

optMain :: OptReader a

reader for this option

optProps :: OptProperties

properties of this option

data OptName

Constructors

OptShort !Char

OptLong !String

data OptReader a

An OptReader defines whether an option matches an command line argument.

Constructors

OptReader [OptName] (CReader a)
option reader

FlagReader [OptName] !a
flag reader

ArgReader (CReader a)
argument reader

CmdReader [String] (String -> Maybe (ParserInfo a))
command reader

data OptProperties

Specification for an individual parser option.

Constructors

OptProperties

Fields

propVisibility :: OptVisibility

whether this flag is shown is the brief description

propHelp :: String

help text for this option

propMetaVar :: String

metavariable for this option

propShowDefault :: Maybe String

what to show in the help text as the default

data OptVisibility

Visibility of an option in the help text.

Constructors

Internal
does not appear in the help text at all

Hidden
only visible in the full description

Visible
visible both in the full and brief descriptions

data CReader a

Constructors

CReader

Fields

crCompleter :: Completer

crReader :: String -> Maybe a

data Parser a where

A Parser a is an option parser returning a value of type a.

Constructors

NilP :: Maybe a -> Parser a

OptP :: Option a -> Parser a

MultP :: Parser (a -> b) -> Parser a -> Parser b

AltP :: Parser a -> Parser a -> Parser a

BindP :: Parser a -> (a -> Parser b) -> Parser b

newtype ParserM r

Constructors

ParserM

Fields

runParserM :: forall x. (r -> Parser x) -> Parser x

newtype Completer

Constructors

Completer

Fields

runCompleter :: String -> IO [String]

data ParserFailure

Result after a parse error.

Constructors

ParserFailure

Fields

errMessage :: String -> IO String

Function which takes the program name as input and returns an error message

errExitCode :: ExitCode

Exit code to use for this error

data OptHelpInfo

Constructors

OptHelpInfo

Fields

hinfoMulti :: Bool

hinfoDefault :: Bool

data OptTree a

Constructors

Leaf a

MultNode [OptTree a]

AltNode [OptTree a]

fromM :: ParserM a -> Parser a

oneM :: Parser a -> ParserM a

manyM :: Parser a -> ParserM [a]

optVisibility :: Option a -> OptVisibility

optMetaVar :: Option a -> String

optHelp :: Option a -> String

optShowDefault :: Option a -> Maybe String